+++ spyswat/swat_calib/analysis/algorithms/__init__.py
@@ -1,5 +1,6 @@
 from .dds import DDS, DDSCalibration
 from .glue import GLUE
 from .parallel_de import ParallelDE
+from .pso import PSO, PSOCalibration
 
-__all__ = ["DDS", "DDSCalibration", "GLUE", "ParallelDE"]
+__all__ = ["DDS", "DDSCalibration", "GLUE", "ParallelDE", "PSO", "PSOCalibration"]